services/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
